control break programm
closes #8
The template was completely rewritten. Now it contains a group change.

Program flow:
First all state abbrevitations with the corresponding state names are read from the stateabb.txt and then loaded into a table. From the file companies.txt a record with a company is read. For each company the state name is determined with state abbrevitation and the data record of the company is written. If it is a company that does not belong to the industry group of the previously read company or the first one that was read, the headings of the industry group are written to the output file.